+++ test/index.js
@@ -2,3 +2,4 @@ require('./functions');
 require('./logger');
 require('./pg');
 require('./mssql');
+require('./express');